requirements.txt
# ===== الإطار الأساسي =====
streamlit==1.32.0
langchain==0.1.7
langchain-community==0.0.20
langchain-openai==0.0.5

# ===== البيانات، الشبكة والأدوات =====
requests==2.31.0
httpx==0.27.0             # يدعم طلبات HTTP غير المتزامنة
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4             # معالجة البيانات

# ===== التصور المرئي (اختياري) =====
plotly==5.18.0
altair==5.2.0

# ======【هام: مزود نموذج اللغة (LLM)، اختر واحدا فقط】======
# الخيار أ: استخدام واجهة برمجة تطبيقات OpenAI مباشرة
openai ==1.10.0

# الخيار ب: استخدام OpenRouter (من خلال واجهة متوافقة مع OpenAI)
# openai==1.6.1           # نفس مكتبة الخيار أ، قم فقط بتكوين `OPENROUTER_BASE_URL` في ملف `.env`

# الخيار ج: استخدام Groq
groq==0.3.0
# langchain-groq==0.1.0   # تكامل LangChain مع Groq